- fixes, add several character fields - update resourcepath of SaveReader to use relative to the currentdirectory - clean unused code

+++ CharacterEditor/Stat.cs
@@ -190,7 +190,7 @@ namespace CharacterEditor
 			get { return GetStatValue("level"); }
 			set
 			{
-				SetStatValue("level", value);
+				SetStatValue("level", (byte)value);
 				OnPropertyChange("Level");
 			}
 		}
